+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (MC 1.20.1) [core] fix: the PlaceholderOnlyParser should only parse placeholders, and should not parse color tags. (This affects the CommandExecutor in `command_event` module and other similar modules)
+- [command_bundle] feature: now will logging when register/unregister a bundle command in console. (Making it easier for debugging.)
+- [command_bundle] fix: the return value of bundle commands is always 1. (It should be 0 if execute successfully.)